refactoring code for journalNutrition + new folder
