+++ src/components/Dashboard/components/Nutrition/MealTypeDropDown.js
@@ -5,8 +5,8 @@ import DropdownButton from "react-bootstrap/DropdownButton";
 import FormControl from "react-bootstrap/FormControl";
 import InputGroup from "react-bootstrap/InputGroup";
 import { useDispatch, useSelector } from "react-redux";
-import { ReactComponent as LikeIcon } from "../../../assets/LikeIcon.svg";
-import { addFoodToJournal } from "../../../state/slices/userinfo";
+import { ReactComponent as LikeIcon } from "../../../../assets/LikeIcon.svg";
+import { addFoodToJournal } from "../../../../state/slices/userinfo";
 
 const MealTypeDropDown = (props) => {
   const {
